move emoji card into own component
